Extract handling of calculations in separate file
Only a small chunk of the ebay script, nevertheless, it muddies
up that file when its only real concern is about cleaning up
the results from the search API.
