+++ app.py
@@ -1,5 +1,6 @@
 from flask import Flask, escape, render_template, request
 import ebay
+import sell_stats
 
 app = Flask(__name__)
 
@@ -13,8 +14,10 @@ def index():
 def search():
     keyword = request.args.get('keyword')
     results = ebay.search_sold(keyword)
+    stats = sell_stats.generate_stats(results['products'])
 
-    return render_template("search.html", keyword=escape(keyword), results=results)
+    return render_template("search.html",
+                           keyword=escape(keyword), results=results, stats=stats)
 
 
 @app.errorhandler(404)
