+++ tests/test_ebay.py
@@ -2,25 +2,6 @@ import pytest
 import ebay
 
 
-def test_calculate_average_on_empty_list():
-    result = ebay.calculate_average([])
-    assert result == 0.0
-
-
-def test_calculate_average_on_short_list():
-    sales = [3, 8]
-    result = ebay.calculate_average(sales)
-
-    assert result == 5.5
-
-
-def test_calculate_average_without_outliers():
-    sales = [1, 22.75, 8.50, 8000]
-    result = ebay.calculate_average(sales)
-
-    assert result == 15.62
-
-
 def test_shipping_is_zero_on_free():
     shipping = {'shippingType': ['Free']}
     result = ebay.calculate_shipping(shipping)
